requirements.txt
lark-parser
tabulate
requests
matplotlib